s/intptr_t/int64_t/; s/uintptr_t/uint64_t/

+++ src/slist.c
@@ -92,8 +92,8 @@ int slist_cmp( slist_t const *i_list, slist_t const *j_list,
     for ( ; i_node != NULL && j_node != NULL;
           i_node = i_node->next, j_node = j_node->next ) {
       int const cmp = STATIC_CAST( int,
-        REINTERPRET_CAST( intptr_t, i_node->data ) -
-        REINTERPRET_CAST( intptr_t, j_node->data )
+        REINTERPRET_CAST( int64_t, i_node->data ) -
+        REINTERPRET_CAST( int64_t, j_node->data )
       );
       if ( cmp != 0 )
         return cmp;
